Reorganized things to make it more secure :)
I was using php's _feature_ of
function names as strings to parse
urls, that was fun but it could allow
attacks via `website/phpSystemFunction`,
which is in fact not entirely true, because
I sent two params to that function,
so many of the php functions crashed.

The improvement was to import the route functions
inside the Shade class and excecute the params
function only if it's inside that file
(guessed with `get_file_contents` and strpos).